+++ daops/utils/consolidate.py
@@ -1,6 +1,7 @@
 import collections
 import glob
 import os
+import re
 
 import xarray as xr
 from roocs_utils.exceptions import InvalidCollection
@@ -18,6 +19,100 @@ from daops.utils.core import _wrap_sequence
 LOGGER = logging.getLogger(__file__)
 
 
+def to_year(time_string):
+    "Returns the year in a time string as an integer."
+    return int(time_string.split("-")[0])
+
+
+def get_year(value, default):
+    """Gets a year from a datetime string. Defaults to the value of `default`
+    if not defined."""
+    if value:
+        return to_year(value)
+    return default
+
+
+def get_years_from_file(fpath):
+    """Attempts to extract years from a file.
+    First by examining the file name. If that doesn't work then it
+    reads the file contents and looks at the time axis.
+
+    Returns a set of years.
+    """
+    # Try to use filename
+    time_comps = os.path.splitext(os.path.basename(fpath))[0].split("_")[-1].split("-")
+    years = {int(tm[:4]) for tm in time_comps if re.match(r"^\d{4,}", tm)}
+
+    # If a range of years
+    if len(years) > 1:
+        years = set(range(min(years), max(years) + 1))
+
+    # If no years, try reading the file
+    if not years:
+        ds = open_xr_dataset(fpath)
+        if hasattr(ds, "time"):
+            years = {int(yr) for yr in ds.time.dt.year}
+
+    return years
+
+
+def get_files_matching_time_range(time_param, file_paths):
+    """
+    Using the settings in `time_param`, examine each file to see if it contains
+    years that are in the requested range.
+
+    The `time_param` can have three types:
+        1. type: "interval":
+           - defined with "start_time" and "end_time"
+        2. type: "series":
+           - defined with a list of "time_values"
+        3. type: "none":
+           - undefined
+
+    It attempts to filter out files that do not match the selected year.
+    For any file that we cannot do this with, the file will be read by
+    xarray.
+
+    Args:
+        time_param (TimeParameter): time parameter of requested date/times
+        file_paths (list): list of file paths
+    Returns:
+        file_paths (list): filtered list of file paths
+    """
+    # Return all file paths if no time inputs specified
+    if time_param.type == "none":
+        return file_paths
+
+    LOGGER.info(f"Testing {len(file_paths)} files in time range: ...")
+    files_in_time_range = []
+
+    # Handle times differently depending on the type of time parameter
+    if time_param.type == "interval":
+
+        tp_start, tp_end = time_param.get_bounds()
+        req_start_year = get_year(tp_start, default=-99999999)
+        req_end_year = get_year(tp_end, default=999999999)
+
+        # Work through the list of file paths checking if each matches
+        for fpath in file_paths:
+            years = get_years_from_file(fpath)
+            if min(years) <= req_end_year and max(years) >= req_start_year:
+                files_in_time_range.append(fpath)
+
+    elif time_param.type == "series":
+
+        # Get requested years and match to files whose years intersect
+        req_years = {to_year(tm) for tm in time_param.asdict().get("time_values", [])}
+
+        for fpath in file_paths:
+            years = get_years_from_file(fpath)
+            if req_years.intersection(years):
+                files_in_time_range.append(fpath)
+
+    LOGGER.info(f"Kept {len(files_in_time_range)} files")
+    return files_in_time_range
+
+
 def consolidate(collection, **kwargs):
     """
     Finds the file paths relating to each input dataset. If a time range has been supplied then only the files
@@ -31,7 +126,7 @@ def consolidate(collection, **kwargs):
     catalog = None
     time = None
 
-    collection = _wrap_sequence(collection.tuple)
+    collection = _wrap_sequence(collection.value)
 
     if not isinstance(collection[0], FileMapper):
         project = get_project_name(collection[0])
@@ -39,62 +134,25 @@ def consolidate(collection, **kwargs):
 
     filtered_refs = collections.OrderedDict()
 
-    if "time" in kwargs:
-        time = kwargs["time"].asdict()
+    time_param = kwargs.get("time")
 
     for dset in collection:
 
         if not catalog:
-            try:
+            file_paths = dset_to_filepaths(dset, force=True)
 
-                consolidated = dset_to_filepaths(dset, force=True)
+            if time_param:
+                file_paths = get_files_matching_time_range(time_param, file_paths)
 
-                if time:
+            # If no files are matched then raise an exception
+            if len(file_paths) == 0:
+                raise Exception(f"No files found in given time range for {dset}")
 
-                    file_paths = consolidated
-                    LOGGER.info(f"Testing {len(file_paths)} files in time range: ...")
-                    files_in_range = []
-
-                    ds = open_xr_dataset(dset)
-
-                    if time["start_time"] is None:
-                        time["start_time"] = ds.time.values.min().strftime("%Y")
-                    if time["end_time"] is None:
-                        time["end_time"] = ds.time.values.max().strftime("%Y")
-
-                    times = [
-                        int(time["start_time"].split("-")[0]),
-                        int(time["end_time"].split("-")[0]) + 1,
-                    ]
-                    required_years = set(range(*[_ for _ in times]))
-
-                    for i, fpath in enumerate(file_paths):
-
-                        LOGGER.info(f"File {i}: {fpath}")
-
-                        ds = open_xr_dataset(fpath)
-
-                        found_years = {int(_) for _ in ds.time.dt.year}
-
-                        if required_years.intersection(found_years):
-                            files_in_range.append(fpath)
-
-                    LOGGER.info(f"Kept {len(files_in_range)} files")
-                    consolidated = files_in_range[:]
-                    if len(files_in_range) == 0:
-                        raise Exception(
-                            f"No files found in given time range for {dset}"
-                        )
-
-            # catch where "time" attribute cannot be accessed in ds
-            except AttributeError:
-                pass
-
-            filtered_refs[dset] = consolidated
+            filtered_refs[dset] = file_paths
 
         else:
             ds_id = derive_ds_id(dset)
-            result = catalog.search(collection=ds_id, time=time)
+            result = catalog.search(collection=ds_id, time=time_param)
 
             if len(result) == 0:
                 result = catalog.search(collection=ds_id, time=None)